requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
python-dotenv
pydantic
google-search-results
gradio
pypdf
faiss-cpu